Export types, set stricter mypy baseline (#790)
* Export types, set stricter mypy baseline

Lyft recently shipped a change to our internal GitHub code
that included a type error, which we should have been able to catch in CI.
Update Cartography to make this possible:
- Follow https://peps.python.org/pep-0561/ to export the type hints
  already present in the Cartography codebase,
  by adding a `py.typed` file and including it in our distributions.
- In our recent buggy change, the lack of typing on decorators (namely `@timeit`)
  meant that although all the functions involved had type hints,
  because they were decorated with `@timeit` mypy saw them typed as `Any`
  and was not able to warn us of a mismatch of argument types/count.
  So, add type annotations to `@timeit` and set `disallow_untyped_decorators = true`.
- Ratchet up the default mypy configuration to be more strict generally.
  We're not yet ready to set `strict = true` globally,
  and that option is unforunately only settable globally,
  not on a per-module basis
  so we cannot opt modules out.
  Set as many global strictness flags as feasible,
  since we cannot incrementally adopt them in a per-module basis.
  Fix all errors outside intel modules;
  for intel modules, `# type: ignore` if there are only 1 or 2 errors
  or else ignore errors on the entire module.

* Ratchet mypy on modules used by Lyft's GitHub ingestion

Lyft has additional changes planned to its GitHub code in the near term,
so ratchet up the strictness flags to catch any errors.
Note that we leave `disallow_any_generics = false`;
setting it to `true` would be more strict
but it's not clear it's yet worth it to add those types,
vs switching to e.g. TypedDict.

The `GraphJob.run_from_json` and `GraphJob.run_from_json_file`
methods lose the default value of `None` for their `parameters` argument;
this appears to be a breaking change for API callers not passing that argument.
However, the parameters argument is passed through to a dict.update call,
which does not accept `None` as an input,
so any callers not passing that argument would have failed at runtime anyways.

* Loosen typing on `GraphJob` json `file_path`s

The `file_path` parameter is passed directly to `open`,
which supports `typing.Union[str, bytes, os.PathLike]`,
and internally at Lyft we always pass a string for this parameter.

We need to parse out a shortname as a string from the path,
so we can't expand this type to include bytes;
only expand to be a Union of str and Path.

* Remove unneeded use of global stats client

Instead of manually scoping our stat, use `get_stats_client`.